dispserv_x11: move the XRandR regression test to samples/gfx
d402590 put this under gfx/display_servers/test/ as a libcheck suite,
which is the convention libretro-common/test/ uses.  samples/ is the
established home for RetroArch-side regression tests, so move it there
and adopt that idiom instead: one self-contained file, a plain main()
returning 0 or 1, "[pass]" lines and "ALL OK", and a Makefile with the
same shape and SANITIZER hook as the rest of samples/gfx.

Dropping libcheck removes a build dependency and a per-case fork.  The
fork is a real loss -- a crash now takes the whole run down at the
first failing case instead of letting the rest report -- but that is
how every other sample here behaves, and for this test the first case
crashing is a perfectly clear signal.

One deliberate departure from the neighbouring samples: those mirror
the shape of the code under test rather than linking it.  This one
links the real gfx/display_servers/dispserv_x11.c, because that file
refers to only four RetroArch-side symbols (g_x11_dpy, g_x11_win,
strlcpy_retro__, video_monitor_set_refresh_rate), so standing it up
costs almost nothing and the test then actually regresses if the
driver changes.  Nothing links -lXrandr or -lX11: the stub is the
point, since a working display is precisely the environment that
cannot produce the failing queries under test.

Six checks covering the same eleven scenarios as before, with the
XRandR stub, the RetroArch stubs and the cases folded into the single
file.  The stub keeps its two useful properties -- screen crtcs
numbered from 100 and output crtcs from 200, so which array the code
walked is read off the log rather than inferred, and per-allocation
tracking so a double free or leak is attributed to the case that
caused it.

Verified both ways: passes on this tree, plain and under
SANITIZER=address,undefined; pointed at 26b640f^ it segfaults in the
first check.

No behaviour change to dispserv_x11.c.
