--- gfx/display_servers/test/.gitignore
@@ -1,3 +1,0 @@
-test_dispserv_x11
-*.gcda
-*.gcno
